Add uuid4 generator to test GeneralHelper
We use UUID's as record ID's in our database. PostgreSQL handles generating these for us when we insert a record. Some of our models have a requirement that linked item ID's are also populated. For example, a `transaction` expects its `invoice_id` and `licence_id` fields to be populated.

When testing though there are times we want to create a record without also having to create its dependents. So far we have been using hard coded UUID's specified within the tests. But we foresee tests where we may need to create a number of items so managing lists of UUID's will become unmaintainable.

This this change adds a UUID4 generator to our current `GeneralHelper` and updates the existing tests to use it where applicable. We will then have this available for those tests that need valid but meaningless UUIDs in the future.