Add new test transaction helper (#175)
We have a number of places now where we need to insert an existing transaction into the DB as part of a test's setup. Currently, we are using `CreateTransactionService` to do this. Elsewhere in the tests though we have always used a dedicated test helper. And for most of these tests, we only need the transaction record. We don't need the additional invoice and licence records `CreateTransactionService` will create.

So to make things consistent and possibly improve performance a little, this change adds a new `TransactionHelper` for creating transactions and refactors existing tests to use it.