Add Database error plugin (#172)
Currently, any database errors result in a generic `500` response from the API. We see the actual error in the logs and in Errbit. But it's a poor experience for the client system. There are also scenarios where a DB error is expected, for example, if you try and add 2 transactions with the same regime and client ID.

When these occur we want our responses to be more meaningful but we also don't want to pollute the code with duplicate `try/catch` blocks all over the place. So, instead, this change will add a new plugin that will add an `onPreResponse` hook. It will check for database errors and generate meaningful Boom errors for the scenarios it expects. For those it doesn't, it will attempt to return something more meaningful than a generic `500`.