Fix waterundertaker validation on charge
https://trello.com/c/xoHCMKqf

When we first implemented the validation for a charge (the data we send to the rules service to calculate a charge) we simply copied what [version 1](https://github.com/defra/charging-module-api) does.

It has a validation rule which requires `eiucSource` be populated if `compensationCharge` is `true`. For `waterUndertaker`, the rule is simply that it is mandatory.

In fact the business rule is

> `EIUC Source` and `Water Undertaker` are mandatory for a compensation charge only (i.e. only where `Compensation Charge` is set to TRUE)

Our QA @anwarisse1 spotted during the testing of both version 1 and version 2 this oversight in the validation rules.

This change fixes the issue in this version.