Fix unit tests for GitHub CI workflow
All the tests were running fine locally but that was because the env vars `GIT_COMMIT` and `DOCKER_TAG` exist. Stubbing them via Sinon caused an error when the tests were run in an environment where they don't exist.

This change replaces Sinon with a process of first cloning `process.env`, making your changes in each test, and restoring it using the clone afterwards (see https://stackoverflow.com/a/42304479/6117745).

Doing it this way rather than adding the env vars to the CI workflow means the plugin should be more stable as it does not depend on the env vars existing.