Add test-only bill run creator
There are times we need to check the the performance of certain actions on large data sets, for example, how long the generate process takes.

We have external tests that can create bill runs and the associated transactions but they are either limited in how many they can add, or where they can be used. They also come with the overhead of waiting for each transaction request to be sent and completed.

Ideally, we could do with a way of creating a bill run without the overhead of sending requests across the network. This new endpoint does exactly that. When called, it will generate bill runs that feature a mix of invoices and licences, plus ones that will be identified as zero value, deminimis, and minimum charge. We can then assess the performance of viewing, generating, approving, and sending bill runs.