Fix issue with generating bill run
It just comes down to putting the await in the right place. By moving the invoice loop that adds the transactions to a method, using `await` within it but not using `await` when we call the method itself, we get the behaviour we want. The API responds immediately, but the job of creating the transactions still continues running without error and deadlocks in the background.