Fix admin requests to regime paths return 500
> The shotr version! If you are an admin user, regime doesn't get populated so lots of functionality goes 'pop'!

It was spotted during testing that if an attempt is made, for example, to create a bill run for a regime but the client ID used is an `admin` then a `500` response is returned and the request fails.

We've quickly spotted the issue. All `admin` authorised systems are automatically granted access to all regimes. So, when we first implemented authorisations we didn't see the need to link `admin` users to regimes within the database. Our `AuthorisationService` would see an `admin` and respond with ***go!***.

As we have iterated, we've added functionality that determines the regime for paths where its required and adds it to `request.app`, the place recommended by Hapi to store data for the life of a request. But that functionality has mistakenly assumed all requests will be with 'non-admin' authorised system credentials. In these cases we grab the regime as part of the authorisation process.

We have come to the decision that admin users should be setup in the DB the same way as non-admins. The regimes are rarely if ever going to change. And if we did add one, it's likely we'd be knee deep in other changes so adding an authorisation to it for the admin users would be a minor thing. Plus it keeps things consistent and means we can keep our authorisation logic as simple as possible.