Remove is admin user clause in AuthService
We don't need the clause because for the purposes of a route with a `regime` in the path the admin user can be treated as any other. We know we will have seeded the user with authorisation to all regimes.

However, the tests still don't pass because we need to update our `AuthorisedSystemHelper` to reflect the new behaviour for admin users.