Fix not requiring area code in transaction create
Our business rules are that the `areaCode` in a create bill run transaction request has to match a list of known area codes.

The problem is we are only checking this if the property is populated. If it's not part of the request, we don't perform the check.

This change fixes the validation to ensure

- it is a mandatory property in the create transaction request
- it still validates against a known list of areas