Add version info to responses plugin (#157)
This is the second part of providing version information about the app and the docker image being used to a response. The first covered [adding git and docker tags to the image](https://github.com/DEFRA/sroc-charging-module-api/pull/155).

With the access we have, we can't confirm for certain which of our docker images provided a response. We can make assumptions based on when the AWS ECS task definition was last updated. But for a period after the update, there is always some uncertainty.

By building the information into the docker image and providing it in every response we can be certain whatever the circumstances. So, this change adds a new plugin that will handle this new feature.