Fix incorrect transaction validation (#183)
This issue originally came to light when we spotted that one of the [acceptance tests](https://github.com/DEFRA/sroc-cha-acceptance-tests) was failing.

It was an attempt to create a transaction with the following payload

```json
{
  "periodStart": "01-APR-2020",
  "periodEnd": "31-03-2021",
  "credit": false,
  "billableDays": 310,
  "authorisedDays": 365,
  "volume": "25125",
  "source": "Tidal",
  "season": "Summer",
  "loss": "Low",
  "twoPartTariff": false,
  "compensationCharge": false,
  "waterUndertaker": false,
  "regionalChargingArea": "Thames",
  "section126Factor": 1,
  "section127Agreement": true,
  "section130Agreement": false,
  "customerReference": "T1234567890A",
  "lineDescription": "Drains within Littleport & Downham IDB",
  "licenceNumber": "ExtraRgnY_record1",
  "chargePeriod": "01-APR-2020 - 31-MAR-2021",
  "region": "T",
  "areaCode": "AGY3"
}
```

The issue was `"periodEnd": "31-03-2021"`. This was causing an error when we attempted to insert the transaction into the DB

```
ERROR:  date/time field value out of range: "31-03-2021"
LINE 1: SELECT '31-03-2021'::date;
               ^
HINT:  Perhaps you need a different "datestyle" setting.
```

So started many a wasted hour looking at PostgreSQL server config, Knex and Objection config and settings, custom pg connection overrides, [pg types](https://github.com/brianc/node-pg-types). All of it wasted!! 😩🤦

When we drilled into what was happening in the code we spotted the problem. `CalculateChargeTranslator` is validating and converting the text string we get into a perfectly fine UTC Date instance. This and the other values are then used to call the rules service and get a result. The problem is `CalculateChargeService` when it's done was only returning the result. It was not returning any of the other lovely, parsed and validated values in `CalculateChargeTranslator`.

Back in the `CreateTransactionService` we were assigning the result to our instance of `TransactionTranslator`. But we had overlooked that its `periodEnd` was still just the original string sent by the client. To avoid duplication, any validation needed by both the charge and the transaction we had just done in the `CalculateChargeTranslator`. We'd forgotten [Joi](https://github.com/sideway/joi) not only validates, but it also parses and castes values to their proper types! 🤦🤦🤦

We've been lucky up to now and unknowingly been relying on PostgreSQL automatically converting strings to dates. The good news is the solution is simple. We need to return not only result of the rules service, but everything `CalculateChargeTranslator` has validated with **Joi**.

By doing this when the `CreateTransactionService` assigns the result to its instance of `TransactionTranslator`, properties like `periodEnd` will be updated with the validated, parsed, and correctly defaulted values we expect.