Fix missing client Id in create transaction (#184)
In [Version 1 of the API](https://github.com/DEFRA/charging-module-api) we added the client Id feature. Basically, the client Id is used by client systems to give them confidence that we don't create duplicate copies of their transactions.

As part of V1, we also added functionality to return the client Id if provided in the create transaction request. We've forgotten that bit with the port to version 2 so this change fixes it.

We are making some changes to reduce complexity and help maintenance. We're are just going to include the client Id whether it's populated or not. This means

- less logic
- fewer tests
- more consistency
- more performance