Add bill run /status endpoint
https://trello.com/c/XNqxpZto

This is the first step in adding a `GET /v2/{regime}/bill-run/{id}/status` endpoint to the API.

Before a client system can perform certain actions, the bill run needs to be in the correct state. This endpoint will allow a client system to quickly check just the status of the endpoint, rather than use an endpoint like `GET /v2/{regime}/bill-run/{id}` and all the extra data it pulls through.

There are also actions which are not instantaneous like `/generate`. This endpoint will be more suitable for client systems to 'poll' in order to confirm if the action has completed.