Add bill run status service (#187)
https://trello.com/c/XNqxpZto

This is the second step in adding a `GET /v2/{regime}/bill-run/{id}/status` endpoint to the API.

Our convention is to do all work in a service rather than directly in the controller. So, even though it's not the largest of tasks, this change adds a `BillRunStatusService` to the project.