Add bill run status presenter
https://trello.com/c/XNqxpZto

This is the third step in adding a `GET /v2/{regime}/bill-run/{id}/status` endpoint to the API.

Our convention is use a **presenter** to convert data from the service into a response client systems expect. So, even though it's the simplest of conversions, this change adds a `BillRunStatusPresenter` to the project.