Add initial implementation
Unfortunately, we cannot use Joi to test the property in the current schema. It does have a [precision()](https://joi.dev/api/?v=17.4.0#numberprecisionlimit) method but that will round values rather than error.

So, as a first past confirming we can use a custom validator that checks the precision using a regex. Our test will still fail, but now because the error type is wrong.