Update implementation to use Joi
When doing the initial implementation we spotted that we had used Joi in our custom validation for the financial year. Specifically, we create a schema separate from the main one.

This was the key to unlocking using Joi to validate `section126Factor`. As long as we can tell Joi not to convert values during validation `precision()` will error. This is the behaviour we wanted and ensures we get a `ValidationError` to remain consistent with all other validations.