Fix section126Factor in CalculateChargeTranslator (#191)
Our business rules state that some values should only have a maximum of 3 decimal places. When we re-tested [Version 1](https://github.com/defra/charging-module-api) we found it wasn't doing this validation so we double-checked if it is necessary.

In the case of `section126Factor` it does need to be validated to 3 decimal places. The rules service can handle up to 15. But the transaction file we generate and export has a hard limit of 3 decimal places for this field. So, that is why we are forced to limit the level of precision.

`volume` was the other case. It has a limit of 150 characters in the file and no fixed format!. The rules service's 15 decimal limit still applies but we are happy to let it do the validation.

So, this change fixes the validation for `section126Factor`.