Add credit transactions and simplify
First key change is doing away with financial years. All we actually care about is creating invoices in order to test the performance of actions. This is because

- the number of invoices in a bill run impacts performance
- the type of invoices in a bill run impacts performance
- financial years are just a away of separating invoices. We can replicate this more easily by just having more customer references

By doing away with financial years we can directly relate the our loop count with the number of invoices that will be created.

The next change is adding the ability to create a 'credit' transaction as part of a 'mixed' licence. The result will still be a positive standard invoice but at least it has a mix of credit and debit.

Finally to reduce the duplication we could see coming when we add more transaction types, we have broken down the data generators. We have created methods for generating the base data so that our transaction type methods can focus just on what is relevant to them.