Move to just controlling invoices
On review, what we are trying to do is create various types of invoices whilst allowing some control over the number and type.

It does not matter how many licences they have under them as long as we have roughly a 3-to-1 ratio of transactions to invoices in a bill run. This will give us a rough equivalent to what we see when WRLS send us their larger bill runs.

So this change scraps all the logic around clever licence generation. Now each invoice gets 1 licence though its type is still dynamic. But what can be driven in is how many invoices of a particular type shouldbe created.