Fix bill run generate request type (PATCH vs POST) (#193)
Clearly, @cruikshanks was not paying attention when he reviewed and approved [Create Generate Bill Run Summary endpoint](https://github.com/DEFRA/sroc-charging-module-api/pull/151). The endpoint has been added as a `POST` request when it should have been a `PATCH`.

We are trying to stick to a convention of

- `GET` for any request to view data. The request should also result in no action or changes to the data
- `POST` for any request to create a new record
- `PATCH` for any request which requires no payload, and initiates an action in the service

The `../bill-runs/{id}/generate` meets the criteria for a `PATCH` and we should have added it as such. This change fixes the mistake @cruikshanks let through 😱🤦