Fix Invoice deminimis query modifier
We spotted during testing that the deminimis Objection query modifier in the `InvoiceModel` was not behaving as expected.

The scenario was we had added 3 transactions for the same invoice, all with a value of £1.26. The net value is £3.78 so it should be subject to deminimis. However, the `GenerateBillRunService` didn't pick it up because the modifier included the clause `credit_value - debit_value > 0`. In this scenario that would be

```
  ((0 - 378) = -378) > 0
```

By reversing the calculation, the modifier should behave as expected

```
  ((378 - 0) = 378) > 0
```

This change applies the update, and looks at some tests to cover this in future.