Add test to cover minimum charge invoices
After reviewing [generate_bill_run_summary](https://github.com/DEFRA/charging-module-api/blob/main/app/services/generate_bill_run_summary.js) in Version 1 of the API we have concluded we also need to ensure any invoices which have been flagged as 'subject to minimum charge' are not picked up by the `deminimis` modifier.

In version 1, this is done by the order of the checks

- **minimum charge** is checked first. Any found are updated so that their debit total is brought up to £25
- then **deminimis** is checked. Because all minimum charge debits have been altered to be £25 they do not get considered

So, we can infer from this order of events that `deminimis` does not apply to anything marked as 'subject to minimum charge'.

Though we can control the order we do things outside of the `InvoiceModel` we can't do that in the deminimis modifier. This new test will help to ensure when it is used, it only returns invoices considered to be deminimis.