Fix minimum charge issues in generate bill run (#195)
It was found during testing that invoices were being flagged up as minimum charge when they shouldn't have been. This change aims to resolve this:

We add additional checks to the InvoiceModel.minimumCharge modifier to ensure that a minimum credit or debit value does exist;
We test that a minimum charge value is required before attempting to create a minimum charge transaction.
Note that rulesServiceHelper.mockValue has been amended as part of this to help with test -- it now accepts values in pence for consistency (with affected tests changed to reflect this), and it returns a Sinon stub that we can use to .restore() the original behaviour before we stub it again (otherwise, we'd need to Sinon.restore() which would revert all stub behaviour -- not an issue at present but it could be in future if other things are stubbed).

Also, minimumValueAdjustments in GenerateBillRunService was renamed to minimumChargeAdjustments, as "minimum value" isn't the terminology we use elsewhere.