Validate water undertaker is boolean in charge (#197)
The rules around `waterUndertaker` in the create transaction and calculate charge requests are a bit complex. If `compensationCharge` is `true` then `waterUndertaker` must be present and a boolean.

If `compensationCharge` is false, we don't really care. However, it seems the rules service does. Even though it's not needed for the calculation if we send anything other than a boolean value the rules service throws a 500.

That wouldn't be too bad if it also had a useful error message that highlighted `waterUndertaker` as the problem. But it doesn't. 😩

So, we want to avoid our client systems getting any error where the cause is unclear. This change adds validation for `waterUndertaker` so that if present it at least ensures its a boolean.