Fix generate bill run service test
We have now merged the changes from [Fix minimum charge issues in generate bill run](https://github.com/DEFRA/sroc-charging-module-api/pull/195) so we know te modifier is working.

But doh! Just setting the volume is going to do nothing! It's the rules service we need to stub. So, having done that and ensured we get a value back less then £5 the test now works.