Output generate bill run timing (#196)
To help monitor the time taken to generate a bill run, we're adding a feature to log the time it takes.

GenerateBillRunService now takes an optional logger, intended to be the server logger object. If it's present, the service will use the logger's info method to log the time taken to generate the bill run.