Update modifier to use Objection `where()`
Was spotted outside of the PR that we were using `whereRaw()` for clauses that can be handled by Objection. `whereRaw()` is made available for cases where you need to to pass a query to Knex in plain form as Objection cannot convert it.

We've agreed our convention should be we only use `whereRaw()` if we have to. If we can we should be doing everything through o
Objection for consistency.