Make logging optional in generator
It has been highlighted that the generator will also be useful for unit testing. When we need to create a bill run with specific invoice types we can use this to simplify the test setup.

But if that is the case, we won't have a logger as that comes from the Hapi server instance. So, this change makes it optional.