Handle 'app crashing' error
It was spotted in testing that if a request for an unknown invoice type is sent it causes the application to crash.

At first this was confusing (we're using Boom!) but then we understood. Whilst we are within the lifecycle of the Hapi request and doing things synchronously any errors we throw or that are thrown are captured by Hapi which then wraps them as Boom `500` errors (if they are not already boom errors).

However, because we have setup the generator to 'kick-off' the generate process asynchronously and then immediately respond all bets are off. Now any errors that are thrown are outside of the lifecycle and have nowhere to go. Hence, when they happen they crash the app.

By wrapping everything in the `go()` method on a `try catch` we can ensure no errors bubble up and crash the app. We should do the same for the generate bill run process, and wherever we use this design. But for now this stops the generator crashing anything.