Title case loss, season & source values in charge
Whilst testing the `/generate` bill run endpoint we came across a situation where the `source`, `season`, and `loss` values were found to be invalid. Yet the rules service still returned a 'valid' result. Hidden away in the response was a message telling us that the values we had sent through did not match those in the rules service.

After some investigation, we found it was purely the 'case' of the strings that were used causing the problem. The values were fine.

So, we are doing 2 things

- we are updating the [SROC Charging Module API](https://github.com/defra/sroc-charging-module-api) to automatically 'title case' any values sent through
- we are also updating it to check for messages from the rules service and respond appropriately

This change covers the first point; converting the `loss`, `season` and `source` values in the request to title case automatically.