Add test only bill run generator (#199)
There are times we need to check the performance of certain actions on large data sets, for example, how long the generate process takes.

We have external tests that can create bill runs and the associated transactions but they are either limited in how many they can add, or where they can be used. They also come with the overhead of waiting for each transaction request to be sent and completed first.

Ideally, we could do with a way of creating a bill run without this overhead. This new endpoint does exactly that. When called, it will generate bill runs that feature a mix of invoices based on the parameters sent. We can then test and assess viewing, generating, approving, and sending bill runs more easily.

The focus is on testing volumes and generating invoices in a particular state. So, we are not concerned with the accuracy of each field. This means we can speed things up by mocking the rules service rather than calling out to it for each request.

The result is a bill run generator we can use externally by calling the endpoint, or within our unit tests to quickly generate populated bill runs.

> Note. As this is intended for testing only, we have been a 'bit' cavalier with the size of the class, for example, and only adding an integration test. We have agreed, however, should it become a key component we will invest more time in improving the code and providing more test coverage 😁