Migrate invoice update to use static
This change migrates the update to the invoice over to using the static method.

One of the things we needed to do this was the invoice Id to update. Previously, it relied on the return result from the `InvoiceService`. We've solved the problem by updating the `CreateMinimumChargeAdjustmentService` to include the `invoice_id` when it creates the adjustment; it needs to be applied to the same invoice so why not copy it.

This means we can just grab the ID of the adjustment! Problem solved.