Fix update of invoice for minimum charge (#200)
During the bill run generate process if an invoice is subject to minimum charge and either its debit or credit value is below £25 we add an additional transaction to make the value up to £25.

The thing to note is we do this for _both sides_ if there is a credit and a debit value. This means we also need to update the debit and credit counts and values in the invoice and bill run record to reflect the new transactions we created.

In testing, though we have found only one side will be updated, dependent on whichever minimum charge transaction is created last.

The cause is a mix of needing to do things within a transaction and attempting to patch an instance of each record type (bill run, invoice, licence).

Currently, the order of events is

- create the minimum charge transactions
- iterate through the transactions
  - first transaction, call the bill run, invoice and licence services passing in the transaction
  - each service gets a copy of the current record in the DB and updates its values based on the transaction
  - returns an instance of the thing ready for updating
- insert the new transaction
- call `patch()` on each instance returned from the services

The problem is, we are doing this in a database transaction. So, the call to `patch()` after the first transaction has not actually happened. This means, for example, when the invoice service gets a copy of the invoice from the DB for the second transaction it still has the original values. Assuming the first transaction was a credit, on the second pass we only update the debit count and values.

- first instance of `InvoiceModel` has only credit values updated. `patch()` updates everything
- second instance of `InvoiceModel` has only debit values updated but also the original credit values (not updated ones). `patch()` updates everything, overwriting the previous change to the credit values

This change fixes the issue. It drops the use of the services which rely on getting the current instance of a record. Instead, we will use the static model classes to generate an update query. As well as solving the problem this has 2 benefits

- it drops the need for querying the DB for copies of the existing data
- each update query is smaller and focused only on updating the relevant fields

We were also able to reduce the complexity of the `BillRunService` by removing the need to know if the call came from creating a transaction or generating a bill run.