Update generate bill run logging
Update to follow the pattern used in `BillRunGenerator`. All the logic for handling the logging is done in one place which means the service can just cal the logging method and let it handle what to do.

Doing it this way opens the possibility for centralising and reusing the logic in the future. For now, we just have it in 2 places but at least they are consistent.