Add bill run ID to generate bill run time logging (#206)
[Output generate bill run timing](https://github.com/DEFRA/sroc-charging-module-api/pull/196) has already proved to be a massive help in understanding how long the generate bill run process takes. But we did spot one more tweak that would make it perfect; including the bill run ID in the log output.

* Note

We updated the logging to follow the pattern used in `BillRunGenerator`. All the logic for handling the logging is done in one place which means the service can just call the logging method and let it handle what to do.

Doing it this way opens the possibility for centralising and reusing the logic in the future. For now, we just have it in 2 places but at least they are consistent.