Add error handling to GenerateBillRunService
If an error occurs during the 'generate bill run' process it will cause the app to crash.

We first noticed this kind of behaviour when working on [Add test only bill run generator](https://github.com/DEFRA/sroc-charging-module-api/pull/199)

> At first [the error] was confusing (we're using Boom!) but then we understood. Whilst we are within the lifecycle of the Hapi request and doing things synchronously any errors we throw or that are thrown are captured by Hapi which then wraps them as Boom `500` errors (if they are not already boom errors).
>
> However, because we have setup the generator to 'kick-off' the generate process asynchronously and then immediately respond all bets are off. Now any errors that are thrown are outside of the lifecycle have nowhere to go. Hence, when they happen they crash the app.

By wrapping everything in the `go()` method of the service in a `try catch` we can ensure no errors bubble up and crash the app. This solved the problem for the `BillRunGenerator` so this change does the same to the `GenerateBillRunService`.