Add error logging and try-catch to service
Again, the `BillRunGenerator` has already done this so we just crib the logic from there.

With both `_calculateAndLogTime()` and `_logError()` being duplicated and used in 2 places, we probably now have a reasonable excuse to move logging into its own 'thing'. But that is beyond the scope for this change so we'll do that in a subsequent PR.