Add auto prefix to BillRunGenerator logging
With the log messages being the same, and the fact the 2 process are likely to be used at the same times it wouldbe useful to disambiguate them.

We'll need to consider this when we refactor out the logging. But for now just a quick change of the error message will suffice.