Minimum charge invoice flag (#205)
According to the draft API docs, viewing the bill run should return a minimumChargeInvoice flag at the invoice level, which isn't currently included in ViewBillRunService. This change amends GenerateBillRunService to implement it.

Note that it does this a little differently to the other invoice-level flags, zeroValueInvoice and deminimisInvoice, which are set by querying the db to return the relevant invoices and patching them to set the flag. For minimumChargeInvoice, we set this invoice-level flag when we update the invoice figures as part of the process of saving minimum charge adjustment transactions to the db.