Allow period start to be same as period end date
https://trello.com/c/Xz1R1zEf

It was noted when reviewing some data from WRLS recently that some charges have the same start and end period dates. Discussions with the WRLS team and the business confirmed this does happen and is acceptable.

So, this change alters the validation so that the `periodStartDate` when creating a transaction or calculating a charge can be the same (not just less than) the `periodEndDate`.