Add stand-in view bill run transaction endpoint (#211)
https://trello.com/c/0m5idNGk

This is the first step in adding a `GET /v2/{regime}/bill-run/{id}/transactions/{id}` endpoint to the API.

Client systems need to see all the details we hold for a transaction. This change adds the endpoint that will be used, plus a basic test. But none of the logic to actually find and present the transaction is yet included.