Lean into docker hub auto build features
First thing we spotted is that the automated build on Docker Hub already makes available a number of environment variables with the details we need, so we don't need to generate them ourselves.

We also want to replicate the process we have in place internally. When we build an image we tag it both as `latest` and with the reference to the git commit (a simpler version rather than the full hash). By adding another hook we are hoping we can do that.